Split data module
